test(app): extract parser helpers to satisfy max-lines
